requirements.txt
Flask==2.3.2
Flask-Login
flask-bcrypt
python-dotenv==1.0.0
pymongo==4.3.3
dnspython==2.3.0
Pillow==9.5.0
python-dotenv
dnspython